pyproject.toml
[project]
name = "python-wp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
